requirements.txt
flask
python-docx
smtplib
email-validator